Updated Session Service for creation of sessions. (Session summary and exercise results)

+++ src/main/java/com/xclr8/api/model/nestedModel/ExerciseResult.java
@@ -13,5 +13,5 @@ public class ExerciseResult {
     double maxAngle;
     double speed;
     //Duration saved in seconds
-    int duration;
+    double duration;
 }
